book-catalog-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pymongo==4.6.0
pydantic==1.10.17
python-dotenv==1.0.1
